Move homemade StrEnum to xdsl.utils.str_enum
