+++ xdsl/ir/core.py
@@ -4,7 +4,6 @@ import re
 from abc import ABC, abstractmethod
 from collections.abc import Iterable, Iterator, Mapping, Sequence
 from dataclasses import dataclass, field
-from enum import Enum
 from io import StringIO
 from itertools import chain
 from typing import (
@@ -27,6 +26,7 @@ from typing_extensions import Self
 from xdsl.traits import IsTerminator, NoTerminator, OpTrait, OpTraitInvT
 from xdsl.utils import lexer
 from xdsl.utils.exceptions import VerifyException
+from xdsl.utils.str_enum import StrEnum
 
 # Used for cyclic dependencies in type hints
 if TYPE_CHECKING:
@@ -460,21 +460,6 @@ class Data(Generic[DataElement], Attribute, ABC):
     @abstractmethod
     def print_parameter(self, printer: Printer) -> None:
         """Print the attribute parameter."""
-
-
-class StrEnum(str, Enum):
-    """
-    Homemade StrEnum. StrEnum is standard in Python>=3.11.
-    """
-
-    @staticmethod
-    def _generate_next_value_(
-        name: str, start: int, count: int, last_values: list[Any]
-    ):
-        return name.lower()
-
-    def __str__(self) -> str:
-        return self.value
 
 
 EnumType = TypeVar("EnumType", bound=StrEnum)
